app/nyc_taxi/client/requirements.txt
kafka-python
python-snappy
lz4
numpy
pandas